requirements.txt
asgiref==3.7.2
auxlib==0.0.43
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.2
djangorestframework==3.15.2
idna==3.10
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
